PRODIGY_BD_04/config.py
import os 
from dotenv import load_dotenv

load_dotenv()

class Config:
    # Flask
    SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret-key")

    # JWT
    JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "jwt-dev-secret")

    #MongoDB
    MONGO_URI = os.getenv("MONGO_URI", "mongodb://localhost:27017/prodigy_bd_04")

    #Redis
    REDIS_HOST = os.getenv("REDIS_HOST", "localhost")
    REDIS_PORT = int(os.getenv("REDIS_PORT", 6379))
    REDIS_DB = int(os.getenv("REDIS_DB", 0))
    CACHE_TTL = int(os.getenv("CACHE_TTL", 60)) #seconds