PRODIGY_BD_04/requirements.txt
Flask
mongoengine
Flask-jwt-extended
bcrypt 
python-dotenv
pymongo
redis